// packages/contracts/typechain-types/contracts/PilotGroth16Verifier.ts
export declare namespace PilotGroth16Verifier {
  export type VerificationKeyStruct = {
    alpha: [BigNumberish, BigNumberish];
    beta: [[BigNumberish, BigNumberish], [BigNumberish, BigNumberish]];
    gamma: [[BigNumberish, BigNumberish], [BigNumberish, BigNumberish]];
    delta: [[BigNumberish, BigNumberish], [BigNumberish, BigNumberish]];
    ic: [BigNumberish, BigNumberish][];
  };

  export type VerificationKeyStructOutput = [
    alpha: [bigint, bigint],
    beta: [[bigint, bigint], [bigint, bigint]],
    gamma: [[bigint, bigint], [bigint, bigint]],
    delta: [[bigint, bigint], [bigint, bigint]],
    ic: [bigint, bigint][]
  ] & {
    alpha: [bigint, bigint];
    beta: [[bigint, bigint], [bigint, bigint]];
    gamma: [[bigint, bigint], [bigint, bigint]];
    delta: [[bigint, bigint], [bigint, bigint]];
    ic: [bigint, bigint][];
  };
}

export interface PilotGroth16VerifierInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "artifactManifestDigest"
      | "assurance"
      | "proofProfile"
      | "verificationKeyCommitment"
      | "verifyProof"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "artifactManifestDigest",
    values?: undefined
  ): string;
  encodeFunctionData(functionFragment: "assurance", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "proofProfile",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "verificationKeyCommitment",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "verifyProof",
    values: [
      [BigNumberish, BigNumberish],
      [[BigNumberish, BigNumberish], [BigNumberish, BigNumberish]],
      [BigNumberish, BigNumberish],
      BigNumberish[]
    ]
  ): string;

  decodeFunctionResult(
    functionFragment: "artifactManifestDigest",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "assurance", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "proofProfile",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "verificationKeyCommitment",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "verifyProof",
    data: BytesLike
  ): Result;
}

export interface PilotGroth16Verifier extends BaseContract {
  connect(runner?: ContractRunner | null): PilotGroth16Verifier;
  waitForDeployment(): Promise<this>;

  interface: PilotGroth16VerifierInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  artifactManifestDigest: TypedContractMethod<[], [string], "view">;

  assurance: TypedContractMethod<[], [string], "view">;

  proofProfile: TypedContractMethod<[], [string], "view">;

  verificationKeyCommitment: TypedContractMethod<[], [string], "view">;

  verifyProof: TypedContractMethod<
    [
      a: [BigNumberish, BigNumberish],
      b: [[BigNumberish, BigNumberish], [BigNumberish, BigNumberish]],
      c: [BigNumberish, BigNumberish],
      signals: BigNumberish[]
    ],
    [boolean],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "artifactManifestDigest"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "assurance"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "proofProfile"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "verificationKeyCommitment"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "verifyProof"
  ): TypedContractMethod<
    [
      a: [BigNumberish, BigNumberish],
      b: [[BigNumberish, BigNumberish], [BigNumberish, BigNumberish]],
      c: [BigNumberish, BigNumberish],
      signals: BigNumberish[]
    ],
    [boolean],
    "view"
  >;

  filters: {};
}